Refactor: 파라미터명 변경 및 클래스 경로 변경
